Input Logic Updated : Pressed, Down, Up
+ Added ProcessEvents method in Window and moved call to glfwPollEvents from Update to it.
~ Input class updated to now keep track of the status of each key and button.
- remove from KeyEventPressed repeat value maybe add repeat event?
+ Added input logic for Pressed, Down, and Up for Keyboard  keys and Mouse buttons. Down is when hold down, Up is when just got released.
+ Added to C# the Inputs also with ScriptGlue in C++ for the new Internal calls.
~ Updated calls to KeyPressed to be changed to key Down where it made sense.
~ Updated logic when reloading C# assembly, will no longer reload when stopping from play mode. Will not reload on play mode and crash, will queue the reload for when play mode stops.

+++ Hazelnut/src/Panels/SceneHierarchyPanel.cpp
@@ -632,7 +632,7 @@ namespace Hazel
 							char buffer[2] = { static_cast<char>(data) };
 							if (ImGui::InputText(name.c_str(), buffer, sizeof(buffer)))
 							{
-								data = *buffer;
+								data = static_cast<uint16_t>(static_cast<uint8_t>(*buffer));
 								scriptInstance->SetFieldValue(name, data);
 							}
 							break;
@@ -819,7 +819,7 @@ namespace Hazel
 							char buffer[2] = { static_cast<char>(data) };
 							if (ImGui::InputText(name.c_str(), buffer, sizeof(buffer)))
 							{
-								data = *buffer;
+								data = static_cast<uint16_t>(static_cast<uint8_t>(*buffer));
 								scriptField.SetValue(data);
 							}
 							break;
@@ -992,7 +992,7 @@ namespace Hazel
 							{
 								auto& scriptFieldInstance = entityFields[name];
 								scriptFieldInstance.Field = field;
-								data = *buffer;
+								data = static_cast<uint16_t>(static_cast<uint8_t>(*buffer));
 								scriptFieldInstance.SetValue(data);
 							}
 							break;
